feat: use tracker endpoint for all tracker objects
