+++ src/data/Dhis2Import.ts
@@ -1,106 +1,123 @@
 import _ from "lodash";
-import {
-    SynchronizationResult,
-    SynchronizationStats,
-    SynchronizationStatus,
-} from "../domain/entities/SynchronizationResult";
+import { SynchronizationResult, SynchronizationStats } from "../domain/entities/SynchronizationResult";
 import i18n from "../locales";
+import { D2Api } from "@eyeseetea/d2-api/2.33";
+import { Ref } from "../domain/entities/ReferenceObject";
 
 export type Status = "OK" | "ERROR";
 
-export interface ImportPostResponse {
+type ImportStats = {
+    created: number;
+    deleted: number;
+    ignored: number;
+    updated: number;
+    total: number;
+};
+
+export type ImportPostResponse = {
     status: Status;
-    message?: string;
-    response?: {
-        status: SynchronizationStatus;
-        imported: number;
-        updated: number;
-        deleted: number;
-        ignored: number;
-        total: number;
-        importSummaries?: Array<{
-            responseType: "ImportSummary";
-            description?: string;
-            status: SynchronizationStatus;
-            href?: string;
-            importCount: {
-                imported: number;
-                updated: number;
-                deleted: number;
-                ignored: number;
-            };
-            reference?: string;
-            conflicts?: {
-                object: string;
-                value: string;
-            }[];
-            // Only for TEI import
-            enrollments?: ImportPostResponse["response"];
-        }>;
+    response: Ref;
+};
+
+export type ImportReportResponse = {
+    status: Status;
+    validationReport: {
+        errorReports: [
+            {
+                uid: string;
+                message: string;
+            }
+        ];
     };
-}
+    stats: ImportStats;
+    bundleReport?: {
+        status: Status;
+        stats: ImportStats;
+        typeReportMap: {
+            [type: string]: {
+                trackerType: string;
+                stats: ImportStats;
+                objectReports: {
+                    uid: string;
+                };
+            };
+        };
+    };
+};
 
 export function processImportResponse(options: {
     title: string;
     model: string;
-    importResult: ImportPostResponse;
+    importResult: ImportReportResponse;
     splitStatsList: boolean;
 }): SynchronizationResult {
     const { title, model, importResult, splitStatsList } = options;
-    const { message, response } = importResult;
-    const status = response ? response.status : "ERROR";
 
-    if (!response) return { title, status, message, rawResponse: importResult };
+    const { bundleReport, status, validationReport } = importResult;
+    const message = status === "OK" ? "Import was successful" : "Import failed";
 
-    // Add inner import summaries
-    const importSummaries = _(response.importSummaries)
-        .flatMap(importSummary => [importSummary, ...(importSummary.enrollments?.importSummaries || [])])
-        .value();
-
-    const aggregatedStatus =
-        _(importSummaries)
-            .map(summary => summary.status)
-            .find(status => status !== "SUCCESS") || status;
-
-    const errors =
-        _.flatMap(
-            importSummaries,
-            ({ reference = "", description, conflicts }) =>
-                conflicts?.map(({ object, value }) => ({
-                    id: reference,
-                    message: _([description, object, value]).compact().join(" "),
-                    details: "",
-                })) ?? (description ? [{ id: reference, message: description, details: "" }] : [])
-        ) ?? [];
-
-    const fields = ["imported", "updated", "ignored", "deleted", "total"] as const;
-    const totalStats: SynchronizationStats = { type: "TOTAL", ..._.pick(response, fields) };
-
-    const eventStatsList = (response.importSummaries || []).map((importSummary): SynchronizationStats => {
+    if (!bundleReport) {
         return {
-            type: i18n.t(`${model} ${importSummary.reference || "-"}`),
-            ...importSummary.importCount,
+            title: title,
+            status: status === "OK" ? "SUCCESS" : "ERROR",
+            message: message,
+            rawResponse: importResult,
         };
-    });
+    }
+
+    const errors = validationReport.errorReports.map(errorReport => ({
+        id: errorReport.uid,
+        message: errorReport.message,
+        details: "",
+    }));
+
+    const totalStats: SynchronizationStats = {
+        type: "TOTAL",
+        imported: bundleReport.stats.created,
+        ...bundleReport.stats,
+    };
+
+    const eventStatsList = _(bundleReport.typeReportMap)
+        .map(value => ({
+            type: i18n.t(`${model}s`),
+            imported: value.stats.created,
+            ...value.stats,
+        }))
+        .reject(value => value.total === 0)
+        .value();
 
     const stats = splitStatsList
         ? _.compact([eventStatsList.length === 1 ? null : totalStats, ...eventStatsList])
         : [totalStats];
 
-    return { title, status: aggregatedStatus, message, errors, stats, rawResponse: importResult };
+    return {
+        title: title,
+        status: status === "OK" ? "SUCCESS" : "ERROR",
+        message: message,
+        errors: errors,
+        stats: stats,
+        rawResponse: importResult,
+    };
 }
 
 export async function postImport(
+    api: D2Api,
     postFn: () => Promise<ImportPostResponse>,
     options: { title: string; model: string; splitStatsList: boolean }
 ): Promise<SynchronizationResult> {
     const { title, model, splitStatsList } = options;
+
     try {
         const response = await postFn();
+        const { response: trackerImportResponse } = response;
+        const trackerJobReport = await api
+            .get<ImportReportResponse>(`/tracker/jobs/${trackerImportResponse.id}/report`)
+            .getData();
+
         return processImportResponse({
             title,
             model: model,
-            importResult: response,
+            importResult: trackerJobReport,
             splitStatsList,
         });
     } catch (error: any) {
